requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Pillow==10.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
email-validator==2.1.0
markdown==3.5.1
bleach==6.0.0
python-slugify==8.0.1